split out dom-utils
